% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_circle.R
\name{create_circle}
\alias{create_circle}
\title{Create the raw data needed to plot a circle}
\usage{
create_circle(radius, res = 100)
}
\arguments{
\item{radius}{A number}

\item{res}{A number}
}
\value{
A dataframe with xy coordinates that will plot a circle
}
\description{
Create the raw data needed to plot a circle
}
