% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_mosaic.R
\name{generate_mosaic}
\alias{generate_mosaic}
\title{A function that takes in user specificiations and returns generative "mosaic" art work}
\usage{
generate_mosaic(
  num_boxes = 1000,
  spiral_color = "random",
  background_color = "random"
)
}
\arguments{
\item{num_boxes}{number specifying number of points}

\item{spiral_color}{input of str type specifying color of inward spiral}

\item{background_color}{input of str type specifying color of background}
}
\value{
A beautiful piece of art :)
}
\description{
A function that takes in user specificiations and returns generative "mosaic" art work
}
