% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_squares.R
\name{interpolate_squares}
\alias{interpolate_squares}
\title{Generates data for the 'Interpolate Squares' visualisation}
\usage{
interpolate_squares(seed = 1, n = 25, ease_function = "")
}
\arguments{
\item{seed}{Seed number}

\item{n}{Number of squares}

\item{ease_function}{Either randomly selected or specified for all squares.
Options: "linear", "quadratic-in", "quadratic-out", "quadratic-in-out", "cubic-in", "cubic-out", "cubic-in-out",
"quartic-in", "quartic-out", "quartic-in-out", "quintic-in", "quintic-out", "quintic-in-out", "sine-in",
"sine-out", "sine-in-out", "circular-in", "circular-out", "circular-in-out", "exponential-in",
"exponential-out", "exponential-in-out", "elastic-in", "elastic-out", "elastic-in-out", "back-in",
"back-out", "back-in-out", "bounce-in", "bounce-out", "bounce-in-out"}
}
\description{
For each square, interpolates the four edges of the square to random line segments whose endpoints lie inside the square
}
\examples{
interpolate_squares()
}
\keyword{interpolate}
