% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procedural_ring.R
\name{procedural_ring}
\alias{procedural_ring}
\title{Procedural Ring}
\usage{
procedural_ring(
  seed = 3,
  n = 400,
  n_circles = 4,
  r_0 = 5000,
  eps = 500,
  eps_big = 5000,
  p_defect = 0.1,
  delta = 2.02
)
}
\arguments{
\item{seed}{The seed number}

\item{n}{The number of steps to make around the circumference of the circle}

\item{n_circles}{The number of times to repeat the algorithm (results are overlaid and interpolated)}

\item{r_0}{The radius of the circle}

\item{eps}{The base amount to offset from the circumference of the circle}

\item{eps_big}{If there is a 'defect', the point jumps to this large radius + eps + noise}

\item{p_defect}{The probability of a defect (an unusualy high offset)}

\item{delta}{The angle (in degrees) between successive radial movements around the circumference (+/-1 deg is added)}
}
\description{
Generates data from a contrived algorithm in which a point continuously moves around
the circumference of a circle and offsets randomly from the circle boundary based on
user-set parameters such as eps.
}
\examples{
procedural_ring()
}
\keyword{algorithm}
\keyword{procedural}
\keyword{ring}
