% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_lines.R
\name{render_lines}
\alias{render_lines}
\title{Render xy coordinates using lines}
\usage{
render_lines(dat, lineAngle = 0, seed = 10, ...)
}
\arguments{
\item{dat}{A dataframe}

\item{lineAngle}{A number}

\item{seed}{A number}
}
\value{
A plot of the rendered dat dataframe
}
\description{
Render xy coordinates using lines
}
