% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shards.R
\name{shards}
\alias{shards}
\title{Function that creates generative art based on user input}
\usage{
shards(iterations, min, max)
}
\arguments{
\item{iterations}{number of figures to create}

\item{min}{minimum value for canvas}

\item{max}{maximum value for the canvas}
}
\value{
generative art
}
\description{
Function that creates generative art based on user input
}
